chore: refactor io_context. use async submission for everything instead of providing the option to disable kernel thread. This makes the code simpler to reason about and maintain.
