+++ io_context/CMakeLists.txt
@@ -10,6 +10,6 @@ include_directories(${LIBURING_INCLUDE_DIRS})
 link_directories(${LIBURING_LIBRARY_DIRS})
 add_definitions(${LIBURING_CFLAGS_OTHER})
 
-add_library(io_context_lib io_uring_ctx.h io_context.h memory_stream_context.h memory_stream_context.cpp)
+add_library(io_context_lib uring_context.h io_context.h memory_stream_context.h memory_stream_context.cpp)
 # need to link fmt because of spdlog
 target_link_libraries(io_context_lib PRIVATE ${LIBURING_LIBRARIES} async_simple::async_simple_static fmt::fmt spdlog::spdlog)
