+++ tests/benchmark.go
@@ -53,7 +53,7 @@ func clientSession(clientID, numMessages int, host string, port int, wg *sync.Wa
 }
 
 func main() {
-	numClients := 900
+	numClients := 1000
 	messagesPerClient := 100
 	host := "127.0.0.1"
 	port := 8080
